Added code to edit listener position and orientation. Listener editing not implemented yet due to error. Fixed error where sound file path text did not clear when switching soundproducers. Added Instructions section to README.md

+++ include/soundproducer.h
@@ -31,11 +31,11 @@ class SoundProducer
 		std::string GetNameString();
 		
 		void setPositionX(double& x); //set x position of sound producer
-		float getPositionX(); //get x position of sound producer
+		double getPositionX(); //get x position of sound producer
 		void setPositionY(double& y); //set y position of sound producer
-		float getPositionY(); //get y position of sound producer
+		double getPositionY(); //get y position of sound producer
 		void setPositionZ(double& z); //set z position of sound producer
-		float getPositionZ(); //get z position of sound producer
+		double getPositionZ(); //get z position of sound producer
 		
 		void InitSoundProducer(std::string& thisName, std::string& filepath, ALuint& buffer,
 								double& x, double& y, double& z);
